// apps/quickview/ultron-api/api.ts
export const ActionDtoTypeEnum = {
    Fold: 'fold',
    Call: 'call',
    Raise: 'raise',
    Check: 'check'
} as const;

export type ActionDtoTypeEnum = typeof ActionDtoTypeEnum[keyof typeof ActionDtoTypeEnum];